+++ R/makeSurvey.R
@@ -2,12 +2,17 @@
 #'
 #' @param doe A design of experiment data frame. Each row is an alternative,
 #' each column is an attribute.
-#' @param nResp Maximum number of survey respondents
-#' @param nAltsPerQ Number of alternatives per question
-#' @param nQPerResp Number of questions per respondent
-#' @param group The name of the group attribute if your doe is designed with a
-#' group (i.e. a labeled experiment design where each set of alternatives
-#' contains one of each of the levels in the `group` attribute.
+#' @param nResp Maximum number of survey respondents.
+#' @param nAltsPerQ Number of alternatives per question.
+#' @param nQPerResp Number of questions per respondent.
+#' @param outsideGood Include an outside good in the choice sets? Defaults to
+#' `FALSE`. If included, the total number of alternatives per question will be
+#' one more than the provided `nAltsPerQ` argument.
+#' @param group The name of the variable to use in a "labeled" design
+#' such that each set of alternatives contains one of each of the levels in
+#' the `group` attribute. If included, the `nAltsPerQ` argument will be
+#' ignored as its value is determined by the unique number of levels in the
+#' `group` variable. Defaults to `NULL`.
 #' @return Returns a conjoint survey by randomly sampling from a design of
 #' experiment data frame. The sampling is done iteratively to remove the
 #' possibility of duplicate alternatives appearing in the same choice question.
@@ -35,20 +40,23 @@
 #'   nQPerResp = 6     # Number of questions per respondent
 #' )
 #' }
-makeSurvey <- function(doe, nResp, nAltsPerQ, nQPerResp, group = NULL) {
+makeSurvey <- function(
+    doe, nResp, nAltsPerQ, nQPerResp, outsideGood = FALSE, group = NULL
+) {
     doe <- as.data.frame(doe) # tibbles break things
     doe$rowID <- seq(nrow(doe)) # Have to set these now to remove dupes later
     if (is.null(group)) {
-        return(randomizeSurvey(doe, nResp, nAltsPerQ, nQPerResp))
+        return(randomizeSurvey(doe, nResp, nAltsPerQ, nQPerResp, outsideGood))
     }
-    return(randomizeSurveyByGroup(doe, nResp, nAltsPerQ, nQPerResp, group))
+    return(randomizeSurveyByGroup(
+        doe, nResp, nAltsPerQ, nQPerResp, outsideGood, group))
 }
 
 repDf <- function(df, n) {
     return(df[rep(seq_len(nrow(df)), n), ])
 }
 
-randomizeSurvey <- function(doe, nResp, nAltsPerQ, nQPerResp) {
+randomizeSurvey <- function(doe, nResp, nAltsPerQ, nQPerResp, outsideGood) {
     survey <- initializeSurvey(doe, nResp, nAltsPerQ, nQPerResp)
     # Randomize rows
     n <- nResp*nAltsPerQ*nQPerResp
@@ -58,10 +66,24 @@ randomizeSurvey <- function(doe, nResp, nAltsPerQ, nQPerResp) {
     survey <- removeDuplicateAlts(survey, nResp, nAltsPerQ, nQPerResp)
     # Re-order column names
     survey <- reorderSurveyCols(survey)
+    if (outsideGood) {
+        survey <- addOutsideGood(survey, nAltsPerQ)
+    }
     return(survey)
 }
 
-randomizeSurveyByGroup <- function(doe, nResp, nAltsPerQ, nQPerResp, group) {
+randomizeSurveyByGroup <- function(
+    doe, nResp, nAltsPerQ, nQPerResp, outsideGood, group
+) {
+    nGroupLevels <- length(unique(doe[,group]))
+    if (nGroupLevels != nAltsPerQ) {
+        warning(paste0(
+            "The nAltsPerQ argument is being set to ", nGroupLevels,
+            " to match the number of unique levels in the ", group, " variable"
+        ))
+        # Over-ride user-provided nAltsPerQ as it is determined by the group
+        nAltsPerQ <- nGroupLevels
+    }
     survey <- initializeSurvey(doe, nResp, nAltsPerQ, nQPerResp, group)
     # Randomize rows by group
     groups <- split(survey, survey[group])
@@ -78,6 +100,9 @@ randomizeSurveyByGroup <- function(doe, nResp, nAltsPerQ, nQPerResp, group) {
     survey <- addMetaData(survey, nResp, nAltsPerQ, nQPerResp)
     # Re-order column names
     survey <- reorderSurveyCols(survey)
+    if (outsideGood) {
+        survey <- addOutsideGood(survey, nAltsPerQ)
+    }
     return(survey)
 }
 
@@ -144,4 +169,40 @@ reorderSurveyCols <- function(survey) {
     survey <- as.data.frame(survey)[,c(metaNames, varNames)]
     survey$rowID <- NULL
     return(survey)
+}
+
+addOutsideGood <- function(survey, nAltsPerQ) {
+    survey <- dummyCodeCheck(survey)
+    # Create outside good rows
+    survey_og <- survey[which(survey$altID == 1),]
+    survey_og[,!names(survey_og)%in%c("respID", "qID", "altID", "obsID")] <- 0
+    survey_og$altID <- nAltsPerQ + 1
+    survey_og$outsideGood <- 1
+    # Add outside good to survey
+    survey$outsideGood <- 0
+    survey <- rbind(survey, survey_og)
+    survey <- survey[order(survey$obsID),]
+    return(survey)
+}
+
+dummyCodeCheck <- function(survey) {
+    types <- getColumnTypes(survey)
+    nonnumeric <- names(types[!types %in% c("integer", "numeric")])
+    if (length(nonnumeric) > 0) {
+        warning(paste0(
+            "The following variables are non-numeric:\n",
+            paste0(nonnumeric, collapse = ", "), "\n\n",
+            "Converting them to dummy-coded variables in order to add ",
+            "outside good"
+        ))
+        survey <- fastDummies::dummy_cols(survey, nonnumeric)
+        survey[,nonnumeric] <- NULL
+    }
+    return(survey)
+}
+
+getColumnTypes <- function(data) {
+  types <- lapply(data, class)
+  test <- function(x) {x[1]}
+  return(unlist(lapply(types, test)))
 }